requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.2
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
magic-filter==1.0.12
multidict==6.2.0
openai==1.69.0
packaging==24.2
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
